requirements.txt
requests
pyephem
